Wish List -  Verify visibility items in the left lower corner
